Fix zero gradient for subtensor assignment.
The variable that is being assigned has its gradient correctly
calculated (g[k]) but later on when the gradient of the variable being
assigned to is calculated g[k] is set to 0. This gives the correct
gradient for the variable being assigned to, but because it shares the
same storage it actually overrides the earlier gradient incorrectly to
zero. This fixes that.